requirements.txt
# API + config
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# Retrieval
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.1

# Generation (HF)
transformers>=4.52.4,<5
accelerate==0.33.0
safetensors==0.4.4

# Utilities
httpx==0.27.0

# Evaluation & Reporting
matplotlib==3.9.2
pandas==2.2.2
tabulate==0.9.0
reportlab==4.2.2
nltk==3.9
rouge-score==0.1.2
bert-score==0.3.13
sacrebleu==2.4.2
protobuf==5.27.2
evaluate==0.4.3

torch==2.6.0
sacremoses==0.1.1